CombMaskMT: fix int overflow

+++ src/filters/comb_mask_mt.zig
@@ -1,13 +1,13 @@
 const std = @import("std");
 const math = std.math;
 
-const vec_len = std.simd.suggestVectorLength(i16) orelse 1;
-const vec_t = @Vector(vec_len, i16);
+const vec_len = std.simd.suggestVectorLength(i32) orelse 1;
+const vec_t = @Vector(vec_len, i32);
 const shift8: vec_t = @splat(8);
 const floor: vec_t = @splat(0);
 const peak: vec_t = @splat(255);
 
-pub fn process(srcp: []const u8, dstp: []u8, stride: u32, width: u32, height: u32, thresinf: i16, thressup: i16) void {
+pub fn process(srcp: []const u8, dstp: []u8, stride: u32, width: u32, height: u32, thresinf: i32, thressup: i32) void {
     const thresinf_v: vec_t = @splat(thresinf);
     const thressup_v: vec_t = @splat(thressup);
 
@@ -28,10 +28,10 @@ pub fn process(srcp: []const u8, dstp: []u8, stride: u32, width: u32, height: u3
                 (@as(vec_t, sd[x..][0..vec_len].*) - @as(vec_t, s[x..][0..vec_len].*));
 
             const sel: vec_t = @select(
-                i16,
+                i32,
                 prod < thresinf_v,
                 floor,
-                @select(i16, prod > thressup_v, peak, prod >> shift8),
+                @select(i32, prod > thressup_v, peak, prod >> shift8),
             );
 
             for (0..vec_len) |i| {
